+++ app/models/variable_form.rb
@@ -1,0 +1,2 @@
+class VariableForm < ActiveRecord::Base
+end