+++ db/migrate/20140314154509_create_forms.rb
@@ -1,0 +1,14 @@
+class CreateForms < ActiveRecord::Migration
+  def change
+    create_table :forms do |t|
+      t.text :folder
+      t.string :name
+      t.string :display_name
+      t.string :code_book
+      t.integer :dataset_id
+
+      t.timestamps
+    end
+    add_index :forms, :dataset_id
+  end
+end