+++ test/models/variable_form_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class VariableFormTest < ActiveSupport::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end