+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added Academic User Group page and added research summary for AUG Members
 - Documentation for repositories can now be synced by system adminstrators directly on the tool or dataset page
 - Datasets and tools now display a list of the contributors
+- Variables now reference the forms on which they were collected
 - Use of Ruby 2.1.1 is now recommended
 - **Gem Changes**
   - Updated to contour 2.4.0